add decorator, update admin

+++ pages/views/admin.py
@@ -1,15 +1,16 @@
-from django.http import HttpResponse, HttpResponseRedirect, StreamingHttpResponse
-from django.template import loader
-from django.contrib.auth.decorators import login_required
-from django.views.decorators.csrf import csrf_exempt
-from pages.views.helper import *
-from authentication.models import *
-from authentication.utils import get_pending_users
-from assign.views import assign
 import csv
-import time
 import datetime
 import json
+import time
+
+from django.core.paginator import Paginator
+from django.http import HttpResponse, HttpResponseRedirect, StreamingHttpResponse
+from django.template import loader
+from django.views.decorators.csrf import csrf_exempt
+
+from assign.views import assign
+from pages.views.helper import *
+from pages.decorators import admin_login_required
 
 
 ADMIN_DIR = 'pages/admin'
@@ -17,43 +18,41 @@ USER_DIR = 'pages/user'
 AUTH_DIR = 'authentication'
 
 
-@login_required
-def index(request):
-    if request.user.is_admin:
-        template = loader.get_template('{}/admin.html'.format(ADMIN_DIR))
-
-        context = {
-            'pending_users': get_pending_users(request.user.group),
-            'approved_users': get_approved_users(request.user.group),
-            'login_user': request.user,
-        }
-        return HttpResponse(template.render(context=context, request=request))
-    else:
-        return HttpResponseRedirect('/')
-
-
-@login_required
-@csrf_exempt
+@admin_login_required
 def modify_users(request):
-    user = get_user_model().objects.get(id=request.POST["id"])
-    if 'approve' in request.POST:
-        user.approve(True)
-    if 'activate' in request.POST:
-        user.activate(True)
-    if 'deactivate' in request.POST:
-        user.activate(False)
-    if 'reject' in request.POST:
-        user.delete()
-    return HttpResponseRedirect('/')
+    if request.method == "POST":
+        user = get_approved_users(request.user.group).get(id=request.POST["id"])
+        if 'approve' in request.POST:
+            user.approve(True)
+        if 'activate' in request.POST:
+            user.activate(True)
+        if 'deactivate' in request.POST:
+            user.activate(False)
+        if 'reject' in request.POST:
+            user.delete()
+        return HttpResponseRedirect('/')
+    else:
+        if request.user.is_admin:
+            template = loader.get_template('{}/admin.html'.format(ADMIN_DIR))
+
+            context = {
+                'pending_users': get_pending_users(request.user.group),
+                'approved_users': get_approved_users(request.user.group),
+                'login_user': request.user,
+            }
+            return HttpResponse(template.render(context=context, request=request))
+        else:
+            return HttpResponseRedirect('/')
 
 
-@login_required
+@admin_login_required
 def dataset(request):
     template = loader.get_template('{}/dataset.html'.format(ADMIN_DIR))
 
     # get all VotingData ids that are not allocated to any user
-    ids = [i.id for i in VotingData.objects.all()]
-    exclude_ids = [i.task.id for i in Assignment.objects.all()]
+    ids = [i.id for i in VotingData.objects.filter(group=request.user.group)]
+    #exclude_ids = [i.task.id for i in Assignment.objects.all()]
+    exclude_ids = []
     for exclude_id in exclude_ids:
         if exclude_id in ids:
             ids.remove(exclude_id)
@@ -61,34 +60,38 @@ def dataset(request):
     # get all VotingData objects and combine them with their respective choices
     voting_data = []
     for id in ids:
-        voting_data.append(VotingData.objects.get(question_id=id))
+        voting_data.append(VotingData.objects.get(id=id))
     for data in voting_data:
-        data.answers = Choice.objects.filter(data_id=data.question_id)
+        data.answers = Choice.objects.filter(data_id=data.id)
 
     # calculate num of data of each type
-    groups = CustomGroup.objects.all()
-    num_data = {}
-    num_data["all"] = FinalizedData.objects.all().count()
-    for group in groups:
-        num_data[group.name] = FinalizedData.objects.all().filter(group=group).count()
+
+    finalized_data = FinalizedData.objects.filter(group=request.user.group)
+    paginator = Paginator(finalized_data, 3)
+
+    page_number = request.GET.get('page')
+    page_obj = paginator.get_page(page_number)
+
+    num_data = finalized_data.count()
 
     context = {
         'title': 'Data Set',
         'num_data': num_data,
+        'page_obj': page_obj,
         'data': voting_data,
-        'types': groups,
         'login_user': request.user,
     }
     return HttpResponse(template.render(context=context))
 
 
-@login_required
+@admin_login_required
 def download_dataset(request):
     """A view that streams a large CSV file."""
     # Generate a sequence of rows. The range is based on the maximum number of
     # rows that can be handled by a single sheet in most spreadsheet
     # applications.
-    rows = ([data.id, data.question_text, data.answer_text, data.type] for data in Data.objects.all())
+    rows = [["id", "question", "answer"]]
+    rows = ([data.id, data.title, data.answer_text] for data in FinalizedData.objects.all())
     pseudo_buffer = Echo()
     writer = csv.writer(pseudo_buffer)
     response = StreamingHttpResponse((writer.writerow(row) for row in rows),
@@ -97,7 +100,8 @@ def download_dataset(request):
     response['Content-Disposition'] = 'attachment; filename=' + filename
     return response
 
-@login_required
+
+@admin_login_required
 @csrf_exempt
 def update(request, question_id):
     if request.method == 'POST':
@@ -111,11 +115,11 @@ def update(request, question_id):
     return HttpResponseRedirect("/dataset")
 
 
-@login_required
+@admin_login_required
 def report(request):
     template = loader.get_template('{}/report.html'.format(ADMIN_DIR))
     i = datetime.datetime.now()
-    users = CustomUser.objects.all()
+    users = get_approved_users(request.user.group)
     groups = compute_group_point()
 
     context = {
@@ -131,9 +135,11 @@ def report(request):
     return HttpResponse(template.render(context=context))
 
 
+@admin_login_required
 def download_report(request):
     """A view that streams a large CSV file."""
-    rows = ([data.id, data.question_text, data.answer_text, data.type] for data in Data.objects.all())
+    rows = [["id", "username", "certificate", "point", "accuracy"]]
+    rows += ([user.id, user.username, user.certificate, user.point, user.accuracy()] for user in get_approved_users(request.user.group))
     pseudo_buffer = Echo()
     writer = csv.writer(pseudo_buffer)
     response = StreamingHttpResponse((writer.writerow(row) for row in rows),
@@ -153,15 +159,17 @@ def log(request):
     return HttpResponse(template.render(context=context))
 
 
+@admin_login_required
 def assign_tasks(request):
-    users = CustomUser.objects.filter(is_active=True, is_approved=True, is_admin=False)
+    users = get_user_model().objects.filter(is_active=True, is_approved=True, is_admin=False)
     assign(users, Assignment, ValidatingData.objects.all(), TaskData)
     assign(users, Assignment, VotingData.objects.filter(is_active=True), TaskData)
     return HttpResponse("Assign Tasks Succeed")
 
 
+@admin_login_required
 def summarize(request):
-    users = get_user_model().objects.all()
+    users = get_approved_users(request.user.group)
     logs = Log.objects.all().filter(checked=False)
     for user in users:
         user_logs = logs.filter(user=user)
